Remove UIImageView Renamed UIZStackedImageView to UIImageView
